Add concrete implementation to sizedcontainer interface; update ArrayTimeSeries/TimeSeries to reflect new inheritance relation
