+++ timeseries/TimeSeries.py
@@ -1,9 +1,10 @@
 import reprlib, numbers, collections
 import math
 from lazy import lazy
+from SizedContainerTimeSeriesInterface import SizedContainerTimeSeriesInterface
 import numpy as np
 
-class TimeSeries:
+class TimeSeries(SizedContainerTimeSeriesInterface):
 	"""
 	Description of time series class
 	
@@ -120,46 +121,6 @@ class TimeSeries:
 					counter += 1
 		return TimeSeries(newValues, newTimes)
 
-	def __repr__(self):
-		#return 'TimeSeries({})'.format([i for i in self._timeseries])
-		"""
-		if len(self._timeseries) > 10:
-			return 'TimeSeries(['+','.join('{}'.format(i) for i in self._timeseries[:5])\
-					+ '...'+','.join('{}'.format(i) for i in self._timeseries[-5:]) + '])'\
-					+ ' -- omitting {} objects'.format(len(self._timeseries) - 10)
-		return 'TimeSeries({})'.format([i for i in self._timeseries]) 
-		"""
-		if len(self._timeseries) > 10:
-			return "TimeSeries: " + str([(t,v) for (t, v) in zip(self._time[:5], self._value[:5])])\
-			+ ".....omitting {} pairs.....".format(len(self._value) - 10) \
-			+ str([(t,v) for (t, v) in zip(self._time[-5:], self._value[-5:])])
-			'''
-			return "TimeSeries" + str([ i for i in self._timeseries[:5]])\
-			+ ".....omitting {} pairs.....".format(len(self._value) - 10) \
-			+ str([ i  for i in self._timeseries[-5:]])
-			'''	
-		return 'TimeSeries: ' + str([(t,v) for (t, v) in zip(self._time, self._value)])
-
-
-	def __str__(self):
-		""" Returns a string represenation of the TimeSeries.
-		If there are more than 100 elements, the rest are abbreviated.
-			
-			Parameters
-			----------
-			None
-
-			Returns
-			-------
-			s : string
-				a string representation of the time series
-		"""
-		if len(self._timeseries) > 10:
-			return "TimeSeries: " + str([(t,v) for (t, v) in zip(self._time[:5], self._value[:5])])\
-			+ ".....omitting {} pairs.....".format(len(self._value) - 10) \
-			+ str([(t,v) for (t, v) in zip(self._time[-5:], self._value[-5:])])
-		return 'TimeSeries: ' + str([(t,v) for (t, v) in zip(self._time, self._value)])
-
 	def __iter__(self):
 		for v in self._value:
 			yield v
@@ -225,14 +186,6 @@ class TimeSeries:
 		if len(self) != len(otherTS) or self._time != otherTS._time:
 			raise ValueError(str(self) + ' and ' + str(otherTS) + ' must have the same time points')
 		return TimeSeries(list(map(lambda t: t[0] * t[1], zip(self._value, otherTS._value))), self._time)
-
-
-	def __abs__(self):
-		return math.sqrt(sum(x * x for x in self))
-
-	# Peilin: not sure how this should behave ???
-	def __bool__(self):
-		return bool(abs(self))
 
 	def __neg__(self):
 		return TimeSeries([-v for v in self._value], self._time)
